chore: move `scan_versions_manifests` into `VersionManifest` `impl`
